// data/github/current/ui/packages/pull-request-viewer/components/__generated__/PullRequestSummaryViewerContentQuery.graphql.ts
export type PullRequestSummaryViewerContentQuery$variables = {
  number: number;
  owner: string;
  repo: string;
};
export type PullRequestSummaryViewerContentQuery$data = {
  readonly repository: {
    readonly pullRequest: {
      readonly " $fragmentSpreads": FragmentRefs<"PullRequestSummaryViewerContent_pullRequest" | "useEmitPageViewEvent_pullRequest">;
    } | null | undefined;
  } | null | undefined;
};
export type PullRequestSummaryViewerContentQuery = {
  response: PullRequestSummaryViewerContentQuery$data;
  variables: PullRequestSummaryViewerContentQuery$variables;
};